backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "ACB technical test"
authors = ["Marcel de Cabanyes <mdecabanyes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.27.2"
responses = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
